# Merging iteration optimization and inclusive condition enhancement.