… source/pic2card/mystique/card_layout/row_column_group.py
… source/pic2card/mystique/extract_properties.py
… source/pic2card/mystique/predict_card.py
… source/pic2card/tests/test_layout.py
… source/pic2card/tests/variables.py