- Optimized the merging containers iteration.
- Made the inclusive logic generic to all element types.
 ![image](https://user-images.githubusercontent.com/43063410/103987051-d6d0c700-51b1-11eb-91d1-0b9c74440661.png)
![image](https://user-images.githubusercontent.com/43063410/103987086-e3551f80-51b1-11eb-86f8-67a32ffd19b8.png)

**Latency changes for the row-column grouping: [Cprofile]**
**Before:** 7394 function calls (7392 primitive calls) in 0.006 seconds
**With current changes:** 1893 function calls (1891 primitive calls) in 0.002 seconds

![image](https://user-images.githubusercontent.com/43063410/103987197-18617200-51b2-11eb-93ae-774b14c5b2d2.png)
![image](https://user-images.githubusercontent.com/43063410/103987236-30d18c80-51b2-11eb-8f5b-c53e7b7d58c8.png)

**Latency changes for the row-column grouping: [Cprofile]**
**Before:** 5371 function calls (5369 primitive calls) in 0.006 seconds 
**With current changes:** 1996 function calls (1991 primitive calls) in 0.002 seconds


- Fixes made in test cases based on the above 2 changes.