@Keerthana786 2021-01-09
> On the latency side, how are the new improvements help? If you have a benchmark to compare previous and current would help.

Updated the latency Cprofiling results in the PR comment section.